refactor: remove unused stuff
