+++ test/integration/env.ts
@@ -1,5 +1,5 @@
-import { readS3Object } from './aws';
 import fs from 'fs';
+import * as AWS from 'aws-sdk';
 
 const outputs: {
   OutputKey: string;
@@ -8,6 +8,22 @@ const outputs: {
 }[] = JSON.parse(process.env.OUTPUT);
 const jwtBucket = process.env.JWT_BUCKET;
 const jwtKey = process.env.JWT_KEY;
+const region = process.env.REGION || 'ap-northeast-1';
+
+AWS.config.update({
+  region
+});
+
+const S3 = new AWS.S3();
+
+export const readS3Object = async (bucket: string, key: string) => {
+  const result = await S3.getObject({
+    Bucket: bucket,
+    Key: key
+  }).promise();
+
+  return result.Body as string;
+};
 
 const main = async () => {
   const endpoint = {
